Address reviewer feedback on #194
Move structures to mod.rs
Remove unused Blkio-related values
Change network stats from total to each eth devices
